src/main/java/com/example/projecttrackingapi/security/DefaultUserDetailsService.java
package com.example.projecttrackingapi.security;

import com.example.projecttrackingapi.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Component;

import java.util.ArrayList;

@Component
@RequiredArgsConstructor
class DefaultUserDetailsService implements UserDetailsService {

    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(final String username) throws UsernameNotFoundException {
        return userRepository.findByUsername(username)
                .map(userEntity -> new User(userEntity.getUsername(), userEntity.getPassword(), new ArrayList<>()))
                .orElseThrow(() -> new UsernameNotFoundException("Username does not exists: " + username));
    }
}
